epsi-liveplotter-main/requirements.txt
keyboard==0.13.5
matplotlib==3.8.4
numpy==2.1.1
PyQt5==5.15.11
PyQt5_sip==12.13.0
pyqtgraph==0.13.7
pyserial==3.5
scipy==1.14.1
